Fix name of perl script Make log more complete Make it easier to rerun without reinitializing

+++ patchwatcher/patchwatcher.sh
@@ -1,11 +1,30 @@
 #!/bin/sh
+# Simple continuous build for Wine 
+# Dan Kegel 2008
+#
+# Watches both git and wine-patches
+# Must do
+#    sudo apt-get install libmail-pop3client-perl 
+# before running first time.
+# Must also set environment variables
+#   PATCHWATCHER_USER=user@host.com
+#   PATCHWATCHER_HOST=mail.host.com
+#   PATCHWATCHER_PASSWORD=userpass 
+# before running.
+
 set -e
 set -x
 
+# Set this to true on first run and after debugging
+initialize=false
+
 TOP=`pwd`
 WORK=$TOP/wine-continuous-workdir
-rm -rf $WORK
-mkdir $WORK
+if $initialize
+then
+    rm -rf $WORK
+fi
+mkdir -p $WORK/mimemail
 
 initialize_tree()
 {
@@ -31,11 +50,13 @@ refresh_tree()
 use_tree()
 {
     cd $WORK
-    if ! perl $TOP/get_next_patch.pl > current.patch 
+    if ! perl $TOP/get-next-patch.pl > current.patch 
     then
        sleep 60
        return 0
     fi
+    echo Processing patch:
+    cat current.patch
 
     rm -rf golden
     mv active golden
@@ -44,17 +65,19 @@ use_tree()
     if ! patch -p1 < ../current.patch > current.log 2>&1
     then
        echo Patch failed
-       # negative report to author
+       # TODO: send negative report to author
     else
        if ! make >> current.log 2>&1
        then
            echo Build failed
-           # negative report to author
+           # TODO: send negative report to author
        else
            echo Patch and build succeeded
-           # positive report to author
+           # TODO: send positive report to author
        fi
+       cat current.log
     fi
+    cd $WORK
     rm -rf active
     mv golden active
 }
@@ -68,6 +91,8 @@ continuous_build()
   done
 }
 
-initialize_tree
-save_tree
+if $initialize
+then
+    initialize_tree
+fi
 continuous_build
